Run InventoryView interactions through inventory-access
InventoryView was changed from class to interface, which use different bytecode instructions for invoking methods. This ABI-breaking change is intended to be resolved by Spigot's "Commodore", but these bytecode modifications are not applied in all cases, such as when using InvUI in a Paper plugin. Moving all interactions with InventoryView to inventory-access fixes this.