Improved slot element supplier
Instead of using Supplier<SlotElement> for gui and inventory ingredients, a custom SlotElementSupplier type now receives the entire requested list of slots and also returns all slot elements in one go. This also allows for an improved GuiSlotElementSupplier, which now keeps the shape of the Gui and can support holes in the structure.